refactor(ARC-1416): move controller and service into correct folder
