+++ api/src/modules/content-page-labels/index.ts
@@ -1,4 +1,4 @@
 export * from './content-page-labels.module';
-export * from './content-page-labels.controller';
-export * from './content-page-labels.service';
+export * from './controllers/content-page-labels.controller';
+export * from './services/content-page-labels.service';
 export * from './content-page-labels.types';
